Create LC-725
We start by creating an array parts of size k to store the resulting linked list parts. Initially, all elements are set to nullptr.

We calculate the length of the input linked list by traversing it using a for loop. This helps us determine the size of each part.

We calculate two important values:

n: The minimum guaranteed size of each part, obtained by dividing the total length by k.
r: The number of extra nodes that need to be distributed among the first r parts, where r is the remainder of the division.
We initialize two pointers, node and prev, to traverse the linked list. node starts at the head of the linked list (root), and prev keeps track of the previous node.

We loop through each part of the result:

Store the current node as the start of the current part in the parts array.
Traverse n + 1 nodes if there are remaining extra nodes (when r > 0). Otherwise, traverse only n nodes.
After traversing, we disconnect the current part from the rest of the list by setting prev->next to nullptr. This ensures that each part is isolated.
Finally, we return the parts array containing the k parts of the linked list.

Complexity:
Time complexity: O(N), where N is the number of nodes in the linked list.
Space complexity:O(K), where k is the number of parts in which the linked list is split.