Create LC-1310.java
Intuition
To solve the problem efficiently, we need to calculate the XOR of subarrays for multiple queries. Calculating the XOR for each subarray directly for every query can be slow and inefficient, especially with the given constraints. Instead, we use a preprocessing step to make the query responses faster.

Approach
Preprocessing with Prefix XOR Array:

We first compute a prefix XOR array pre, where pre[i] is the XOR of all elements from the start of the array up to index i. This allows us to quickly compute the XOR for any subarray.
Compute XOR for Queries:

Given the prefix XOR array, for any query that asks for the XOR of the subarray from index lefti to righti, we can efficiently compute the result using:
pre[righti] if lefti is 0.
pre[righti] ^ pre[lefti - 1] otherwise. This is because pre[righti] gives the XOR from the start of the array up to righti, and subtracting pre[lefti - 1] removes the contribution of the elements before lefti.